requirements.txt
aiosqlite==0.22.1
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
Authlib==1.6.9
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cloudpickle==3.1.2
cryptography==46.0.5
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.1
google-adk==1.27.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-cloud-aiplatform==1.141.0
google-cloud-appengine-logging==1.8.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.40.1
google-cloud-bigquery-storage==2.36.2
google-cloud-bigtable==2.35.0
google-cloud-core==2.5.0
google-cloud-dataplex==2.16.0
google-cloud-discoveryengine==0.13.12
google-cloud-iam==2.21.0
google-cloud-logging==3.14.0
google-cloud-monitoring==2.29.1
google-cloud-pubsub==2.36.0
google-cloud-resource-manager==1.16.0
google-cloud-secret-manager==2.26.0
google-cloud-spanner==3.63.0
google-cloud-speech==2.37.0
google-cloud-storage==3.9.0
google-cloud-trace==1.18.0
google-crc32c==1.8.0
google-genai==1.67.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
graphviz==0.21
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
mcp==1.26.0
mmh3==5.2.1
opentelemetry-api==1.38.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==26.0
proto-plus==1.27.1
protobuf==5.29.6
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
pyOpenSSL==25.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
sqlalchemy-spanner==1.17.2
sqlparse==0.5.5
sse-starlette==3.3.2
starlette==0.52.1
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
watchdog==6.0.0
websockets==15.0.1
zipp==3.23.0
